Safari-611.1.10/LayoutTests/http/tests/performance/performance-resource-timing-cached-entries-expected.txt

CONSOLE MESSAGE: current entries: 0
PASS foundResources.length is 2
